feat: add jsonrpc endpoint to unchained api
