+++ node/coinstacks/bnbsmartchain/api/src/controller.ts
@@ -1,25 +1,10 @@
 import { ethers } from 'ethers'
-import { Body, Controller, Example, Get, Path, Post, Query, Response, Route, Tags } from 'tsoa'
+import { Example, Get, Query, Response, Route, Tags } from 'tsoa'
 import { Blockbook } from '@shapeshiftoss/blockbook'
 import { Logger } from '@shapeshiftoss/logger'
-import {
-  BadRequestError,
-  BaseAPI,
-  BaseInfo,
-  InternalServerError,
-  SendTxBody,
-  ValidationError,
-} from '../../../common/api/src' // unable to import models from a module with tsoa
-import {
-  API,
-  Account,
-  GasFees,
-  Tx,
-  TxHistory,
-  GasEstimate,
-  TokenType,
-  TokenMetadata,
-} from '../../../common/api/src/evm' // unable to import models from a module with tsoa
+import { BaseAPI, InternalServerError, ValidationError } from '../../../common/api/src' // unable to import models from a module with tsoa
+import { API, GasEstimate, GasFees } from '../../../common/api/src/evm' // unable to import models from a module with tsoa
+import { EVM } from '../../../common/api/src/evm/controller'
 import { Service } from '../../../common/api/src/evm/service'
 import { GasOracle } from '../../../common/api/src/evm/gasOracle'
 
@@ -55,138 +40,12 @@ export const service = new Service({
   rpcApiKey: RPC_API_KEY,
 })
 
+// assign service to be used for all instances of EVM
+EVM.service = service
+
 @Route('api/v1')
 @Tags('v1')
-export class BNBSmartChain extends Controller implements BaseAPI, API {
-  /**
-   * Get information about the running coinstack
-   *
-   * @returns {Promise<BaseInfo>} coinstack info
-   */
-  @Example<BaseInfo>({
-    network: 'mainnet',
-  })
-  @Get('info/')
-  async getInfo(): Promise<BaseInfo> {
-    return {
-      network: NETWORK as string,
-    }
-  }
-
-  /**
-   * Get account details by address
-   *
-   * @param {string} pubkey account address
-   *
-   * @returns {Promise<Account>} account details
-   *
-   * @example pubkey "0xC480394241c76F3993ec5D121ce4F198f7844443"
-   */
-  @Example<Account>({
-    balance: '294505451261967226',
-    unconfirmedBalance: '0',
-    nonce: 74,
-    pubkey: '0xC480394241c76F3993ec5D121ce4F198f7844443',
-    tokens: [
-      {
-        balance: '0',
-        contract: '0x04756126F044634C9a0f0E985e60c88a51ACC206',
-        decimals: 18,
-        name: 'Carbon',
-        symbol: 'CSIX',
-        type: 'BEP20',
-      },
-    ],
-  })
-  @Response<BadRequestError>(400, 'Bad Request')
-  @Response<ValidationError>(422, 'Validation Error')
-  @Response<InternalServerError>(500, 'Internal Server Error')
-  @Get('account/{pubkey}')
-  async getAccount(@Path() pubkey: string): Promise<Account> {
-    return service.getAccount(pubkey)
-  }
-
-  /**
-   * Get transaction history by address
-   *
-   * @param {string} pubkey account address
-   * @param {string} [cursor] the cursor returned in previous query (base64 encoded json object with a 'page' property)
-   * @param {number} [pageSize] page size (10 by default)
-   * @param {number} [from] from block number (0 by default)
-   * @param {number} [to] to block number (pending by default)
-   *
-   * @returns {Promise<TxHistory>} transaction history
-   *
-   * @example pubkey "0xC480394241c76F3993ec5D121ce4F198f7844443"
-   */
-  @Example<TxHistory>({
-    pubkey: '0xC480394241c76F3993ec5D121ce4F198f7844443',
-    cursor:
-      'eyJibG9ja2Jvb2tQYWdlIjoxLCJleHBsb3JlclBhZ2UiOjEsImV4cGxvcmVyVHhpZCI6IjB4MTVkODlmMTdmOWQ2NmZiM2VkMTIwZTc3Mjc0NWQ1MTIzODQ3ZGY3MjM2N2EyNWMwNGUzYWRjY2JlYjgwM2FiMyIsImJsb2NrYm9va1R4aWQiOiIweDJkZjdkZTM1ZTk1NmE3ZTY1M2M1YTE3Mjc2NzM5ZGViNmY0NGUzOGZlMDRiMmUxYjcwYTc2ZmE4YjA2NWNkYjIiLCJibG9ja0hlaWdodCI6MjU4Mzk1Njl9',
-    txs: [
-      {
-        txid: '0x025b88d4b35e1fd28ee372deb1cb28c2c862703dce444629c47e27b8b8759cc4',
-        blockHash: '0x695b9e8a01b9564387bde6d52fd2775867c7b56ee0c1a9d61bbcc2b38a9c835f',
-        blockHeight: 25839827,
-        timestamp: 1676923869,
-        status: 1,
-        from: '0xC480394241c76F3993ec5D121ce4F198f7844443',
-        to: '0x215B8E1810Bb8FCcf2D90eE87631F16B5F4a895f',
-        confirmations: 50,
-        value: '1200000000000000000',
-        fee: '105000000000000',
-        gasLimit: '21000',
-        gasUsed: '21000',
-        gasPrice: '5000000000',
-      },
-    ],
-  })
-  @Response<BadRequestError>(400, 'Bad Request')
-  @Response<ValidationError>(422, 'Validation Error')
-  @Response<InternalServerError>(500, 'Internal Server Error')
-  @Get('account/{pubkey}/txs')
-  async getTxHistory(
-    @Path() pubkey: string,
-    @Query() cursor?: string,
-    @Query() pageSize = 10,
-    @Query() from?: number,
-    @Query() to?: number
-  ): Promise<TxHistory> {
-    return service.getTxHistory(pubkey, cursor, pageSize, from, to)
-  }
-
-  /**
-   * Get transaction details
-   *
-   * @param {string} txid transaction hash
-   *
-   * @example txid "0x025b88d4b35e1fd28ee372deb1cb28c2c862703dce444629c47e27b8b8759cc4"
-   *
-   * @returns {Promise<Tx>} transaction payload
-   */
-  @Example<Tx>({
-    txid: '0x025b88d4b35e1fd28ee372deb1cb28c2c862703dce444629c47e27b8b8759cc4',
-    blockHash: '0x695b9e8a01b9564387bde6d52fd2775867c7b56ee0c1a9d61bbcc2b38a9c835f',
-    blockHeight: 25839827,
-    timestamp: 1676923869,
-    status: 1,
-    from: '0xC480394241c76F3993ec5D121ce4F198f7844443',
-    to: '0x215B8E1810Bb8FCcf2D90eE87631F16B5F4a895f',
-    confirmations: 50,
-    value: '1200000000000000000',
-    fee: '105000000000000',
-    gasLimit: '21000',
-    gasUsed: '21000',
-    gasPrice: '5000000000',
-  })
-  @Response<BadRequestError>(400, 'Bad Request')
-  @Response<ValidationError>(422, 'Validation Error')
-  @Response<InternalServerError>(500, 'Internal Server Error')
-  @Get('tx/{txid}')
-  async getTransaction(@Path() txid: string): Promise<Tx> {
-    return service.getTransaction(txid)
-  }
-
+export class BNBSmartChain extends EVM implements BaseAPI, API {
   /**
    * Get the estimated gas cost of a transaction
    *
@@ -244,58 +103,5 @@ export class BNBSmartChain extends Controller implements BaseAPI, API {
   @Get('/gas/fees')
   async getGasFees(): Promise<GasFees> {
     return service.getGasFees()
-  }
-
-  /**
-   * Broadcast signed raw transaction
-   *
-   * @param {SendTxBody} body serialized raw transaction hex
-   *
-   * @returns {Promise<string>} transaction id
-   *
-   * @example body {
-   *    "hex": "0xf86c0a85046c7cfe0083016dea94d1310c1e038bc12865d3d3997275b3e4737c6302880b503be34d9fe80080269fc7eaaa9c21f59adf8ad43ed66cf5ef9ee1c317bd4d32cd65401e7aaca47cfaa0387d79c65b90be6260d09dcfb780f29dd8133b9b1ceb20b83b7e442b4bfc30cb"
-   * }
-   */
-  @Example<string>('0xb9d4ad5408f53eac8627f9ccd840ba8fb3469d55cd9cc2a11c6e049f1eef4edd')
-  @Response<BadRequestError>(400, 'Bad Request')
-  @Response<ValidationError>(422, 'Validation Error')
-  @Response<InternalServerError>(500, 'Internal Server Error')
-  @Post('send/')
-  async sendTx(@Body() body: SendTxBody): Promise<string> {
-    return service.sendTx(body)
-  }
-
-  /**
-  /**
-   * Get token metadata
-   *
-   * @param {string} contract contract address
-   * @param {string} id token identifier
-   * @param {TokenType} type token type (erc721 or erc1155)
-   *
-   * @returns {Promise<TokenMetadata>} token metadata
-   *
-   * @example contractAddress "0x5000Aecf70A935E49C5A1c8AC3469430A8d39072"
-   * @example id "47240"
-   * @example type "erc721"
-   */
-  @Example<TokenMetadata>({
-    name: 'Giant Squid zkGalxe Special NFT | Manta Network',
-    description: '',
-    media: {
-      url: 'https://cdn.galxe.com/galaxy/mantanetwork/54518748-d6e7-4ab4-9386-e941bfc42945.png',
-      type: 'image',
-    },
-  })
-  @Response<ValidationError>(422, 'Validation Error')
-  @Response<InternalServerError>(500, 'Internal Server Error')
-  @Get('/metadata/token')
-  async getTokenMetadata(
-    @Query() contract: string,
-    @Query() id: string,
-    @Query() type: TokenType
-  ): Promise<TokenMetadata> {
-    return service.getTokenMetadata(contract, id, type)
   }
 }
