+++ node/coinstacks/common/api/src/types.ts
@@ -4,17 +4,3 @@
 export interface Cursor {
   page: number
 }
-
-export interface RPCRequest {
-  jsonrpc: '2.0'
-  id: string
-  method: string
-  params: Array<unknown>
-}
-
-export interface RPCResponse {
-  jsonrpc: '2.0'
-  id: string
-  result?: unknown
-  error?: Record<string, unknown>
-}
